requirements.txt
requests
tk
pillow
pyyaml